Rename gen/installer to gen/build_deploy
Once a month or so someone suggests a bunch of new code should go in this folder
or gets confused about this vs. dcos_installer. This name change should help
with that. The folder does _not_ contain code that installs dc/os.
