+++ gen/build_deploy/bash/dcos-launch.spec
