+++ gen/build_deploy/bash/dcos_generate_config.sh.in
