+++ setup.py
@@ -27,7 +27,7 @@ setup(
         'gen',
         'gen.aws',
         'gen.azure',
-        'gen.installer',
+        'gen.build_deploy',
         'pkgpanda',
         'pkgpanda.build',
         'pkgpanda.http',
@@ -91,10 +91,10 @@ setup(
             'azure/templates/acs.json',
             'azure/templates/azure.html',
             'azure/templates/azuredeploy.json',
-            'installer/bash/dcos_generate_config.sh.in',
-            'installer/bash/Dockerfile.in',
-            'installer/bash/installer_internal_wrapper.in',
-            'installer/bash/dcos-launch.spec',
+            'build_deploy/bash/dcos_generate_config.sh.in',
+            'build_deploy/bash/Dockerfile.in',
+            'build_deploy/bash/installer_internal_wrapper.in',
+            'build_deploy/bash/dcos-launch.spec',
             'coreos-aws/cloud-config.yaml',
             'coreos/cloud-config.yaml'
         ] + get_advanced_templates(),
